+++ R/CreateTableOneJS.R
@@ -76,8 +76,10 @@ CreateTableOne2 <- function(data, strata, vars, factorVars, includeNA = F, test 
         res0$CatTable[[i]][[j]]$level <- labeldata[.(j, lvs), val_label]
       }
     }
-    ptb1.rn <- rownames(print(res0, showAllLevels = showAllLevels, printToggle = printToggle, quote = quote, varLabels = Labels, nonnormal = nonnormal,
-                              catDigits = catDigits, contDigits = contDigits))
+    
+    ptb1.res0 <- print(res0, showAllLevels = showAllLevels, printToggle = printToggle, quote = quote, varLabels = Labels, nonnormal = nonnormal,
+                       catDigits = catDigits, contDigits = contDigits)
+    ptb1.rn <- rownames(ptb1.res0)
     ptb1.rn <- gsub("(mean (SD))", "", ptb1.rn, fixed=T)
   }
   
@@ -95,7 +97,7 @@ CreateTableOne2 <- function(data, strata, vars, factorVars, includeNA = F, test 
   rownames(ptb1) <- gsub("(mean (SD))", "", rownames(ptb1), fixed=T)
   if (Labels & !is.null(labeldata)){
     rownames(ptb1) <- ptb1.rn
-    
+    if (showAllLevels == T) ptb1[, 1] <- ptb1.res0[, 1]
   }
   
   #cap.tb1 = paste("Table 1: Stratified by ", strata, sep="")
@@ -182,6 +184,11 @@ CreateTableOneJS <- function(vars, strata = NULL, strata2 = NULL, data, factorVa
   
   if (is.null(strata)){
     
+    if (Labels & !is.null(labeldata)){
+      labelled::var_label(data) <- sapply(names(data), function(v){as.character(labeldata[get("variable") == v, "var_label"][1])}, simplify = F)
+      data.table::setkey(labeldata, variable, level)
+    }
+    
     res <- tableone::CreateTableOne(vars =vars, data = data, factorVars = factorVars, includeNA = includeNA, test = test, 
                          testApprox = testApprox, argsApprox = argsApprox,
                          testExact = testExact, argsExact = argsExact,
@@ -191,8 +198,6 @@ CreateTableOneJS <- function(vars, strata = NULL, strata2 = NULL, data, factorVa
     factor_vars <- res[["MetaData"]][["varFactors"]]
     
     if (Labels & !is.null(labeldata)){
-      labelled::var_label(data) <- sapply(names(data), function(v){as.character(labeldata[get("variable") == v, "var_label"][1])}, simplify = F)
-      data.table::setkey(labeldata, variable, level)
       for (i in seq_along(res$CatTable)){
         for(j in factor_vars){
           lvs <- res$CatTable[[i]][[j]]$level
@@ -262,9 +267,13 @@ CreateTableOneJS <- function(vars, strata = NULL, strata2 = NULL, data, factorVa
           res$CatTable[[i]][[j]]$level <- labeldata[.(j, lvs), val_label]
         }
       }
-      ptb1.rn <- rownames(print(res, showAllLevels = showAllLevels, printToggle = printToggle, quote = quote, varLabels = Labels, nonnormal = nonnormal,
-                                catDigits = catDigits, contDigits = contDigits))
+      
+      ptb1.res <- print(res, showAllLevels = showAllLevels, printToggle = printToggle, quote = quote, varLabels = Labels, nonnormal = nonnormal,
+                        catDigits = catDigits, contDigits = contDigits)
+      ptb1.rn <- rownames(ptb1.res)
       rownames(ptb1.cbind) <- gsub("(mean (SD))", "", ptb1.rn, fixed=T)
+      if (showAllLevels == T) {ptb1.cbind[, 1] <- ptb1.res[, 1]}
+      
       cap.tb1 <- paste("Stratified by ", labeldata[get("variable") == strata, "var_label"][1], "(", paste(unlist(labeldata[get("variable") == strata, "val_label"]), collapse=", "), ") & ", labeldata[get("variable") == strata2, "var_label"][1], sep="")
     }
     
@@ -297,9 +306,12 @@ CreateTableOneJS <- function(vars, strata = NULL, strata2 = NULL, data, factorVa
           res0$CatTable[[i]][[j]]$level <- labeldata[.(j, lvs), val_label]
         }
       }
-      ptb1.rn <- rownames(print(res0, showAllLevels = showAllLevels, printToggle = printToggle, quote = quote, varLabels = Labels, nonnormal = nonnormal,
-                                catDigits = catDigits, contDigits = contDigits))
+      
+      ptb1.res0 <- print(res0, showAllLevels = showAllLevels, printToggle = printToggle, quote = quote, varLabels = Labels, nonnormal = nonnormal,
+                         catDigits = catDigits, contDigits = contDigits)
+      ptb1.rn <- rownames(ptb1.res0)
       ptb1.rn <- gsub("(mean (SD))", "", ptb1.rn, fixed=T)
+      
       #vals.tb1 <- c(NA, unlist(sapply(vars, function(v){labeldata[get("variable") == v, "val_label"]})))
     }
     
@@ -311,6 +323,7 @@ CreateTableOneJS <- function(vars, strata = NULL, strata2 = NULL, data, factorVa
     rownames(ptb1) <- gsub("(mean (SD))", "", rownames(ptb1), fixed=T)
     if (Labels & !is.null(labeldata)){
       rownames(ptb1) <- ptb1.rn
+      if (showAllLevels == T) {ptb1[, 1] <- ptb1.res0[, 1]}
     
       }